pyproject.toml
[project]
name = "irctc-dynamic-router"
version = "1.0.0"
description = "Multi-turn RL environment simulating Indian Railways ticket booking with WL constraints, budget management, and multi-hop journey planning."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Solo Warrior" }
]

dependencies = [
    "openenv-core",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "openai>=1.50.0",
    "requests>=2.32.0",
    "pyyaml>=6.0",
    "websockets>=12.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["server*"]
